requirements.txt
google-generativeai==0.8.5
fastapi==0.115.12
scipy==1.15.2
pydantic==2.11.4
dotenv==0.9.9
uvicorn==0.34.2
TTS==0.22.0
gradio==5.29.0
uuid==1.30
git+https://github.com/Wikidepia/g2p-id
